Create Ayogu Cleopatra Nmesonmachi. Reg no: 2024704083
This Python Program implements a grading system using IF and ELIF statements. It takes the students score as input and prints the corresponding grade based on the following score

• 70 - 100 = A
•  60-69    = B
• 50-59     = C
• 45-49     = D
• 40-44     = E
• 20-39     = F

For each valid score, the Program also 
prints "YES" to Confirm that the grading was successful. If the Score is outside the valid range. It prints "Invalid score" and "No".

Created By: Ayogu Cleopatra Nmesonmachi 
Reg No: 2024704083